*: move [backup_]nexthop from zapi_nhg into nhg_nexthop sub-structure
The zapi nexthop group structure is unable to reference other
nexthop groups. Actually, only nexthop IP addresses can be
associated to a nexthop group via the ZAPI interface.

This first commit is a preliminary commit that prepares the
ZAPI interface in order to host a second type of nexthop group,
by moving the current zapi_nexthop definition into a specific
sub-structure. This commit does not change the functionality.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>