*: add nexthop_group group ids support, part 1
A nexthop group can either be a list of nexthops IP addresses
or a list of nexthop groups. This is not supported yet.

Extend the nexthop group structure to support a list of nexthop
groups. Expose a vty API inside the nexthop group node, to
host a new command to configure a group name. That group
name points to another nexthop group.

This commit only focuses to changes in the ZAPI interface, and on
the sharpd daemon. The handling of group ids on zebra is stopped
in 'zebra_nhg_proto_add()' function.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>