lib, zebra: add nhg group ids support, part 2, add nhg ref
When a nexthop group with dependencies is attached to a route entry,
the real nexthop flags are not easily reachable for read or write
operations.

In order to do so, at the route install, or at the nexthop group
update, for each nexthop dependency, use an internal nhg pointer to
remember the used nexthop group. The pointer is refreshed at nexthop
group replacement.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>