sharpd, zebra, lib: add recursive support for protocol nhid
Nexthops with an undefined interface index are not supported,
whereas recursive routing may be available.

Add the 'allow-recursion' command under nexthop-group, to
autorise nexthops with undefined interface to be sent to
zebra.
Handle the ALLOW_RECURSION flag in nexthop groups, and
dynamically resolve the nexthop with a resolved nexthop,
if needed.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>