bgpd: add nexthop tracking support for nexthop-groups
Nexthop tracking changes should be handled for each BGP
nexthop-group, prior to evaluate each BGP path.

Add a hook on nexthop tracking to parse the list of available
bgp_nhg_cache entries matching the changed nexthop tracking
context.

If the nexthop is invalid, the nexthop-group is detached and
removed from the BGP paths.
If the nexthop is valid, but resolves over a blackhole route,
the paths are detached from the nexthop-group and re-announced
to zebra. Else the nexthop-group is updated to take into
account the new resolved path.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>