bgpd: add nexthop group support for single path routes
Today, BGP does not support PIC as described in the below link. Each
prefix is sent along with its next-hop, and the next-hop association
is done at ZEBRA level.
Basically, when a failover happens on a 600000 route setup, ZEBRA
will need to receive 600000 messages to switch all the routes to the
appropriate nexthop, ... which takes time.

Let us BGP take control over the nexthop by generalising the Nexthop
Group support at the protocol level.

This commit relies on the technical enhancements provided in the
nexthop group library. It reuses the RECURSION, IBGP and SRTE
flags.

It is worth noting that this commit does not apply to multiple paths,
as a separate commit will introduce the nexthop group dependency
functionality.

Link: https://datatracker.ietf.org/doc/draft-ietf-rtgwg-bgp-pic

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>