topotests: add bgp_nhg_topo1 test
This topology uses an IGP network with BGP between r1, and
r5/r6 on the one hand, and two ces at the edge side, on the
other hand. The bgp nexthop-group mode is used.
An identifier is forged by BGP, and is sent to ZEBRA for resolution.
The nexthop groups are checked for usage, including in VRFs.
When IGP changes, the nexhops are refreshed to take into account
the resolution.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>