zebra: add prefix evaluation at NHG updates
When a BGP configuration receives and installs a prefix with multiple
nexthops, if a link down event happens, the nexthop resolution is not
propagated. The below example illustrates the 10.0.5.1 prefix, where
the resolution is given by BGP.

> # show ip nht
> [..]
> 10.0.5.1
>  resolved via bgp
>  via 10.0.7.2, r1-r2-eth2 (vrf default), weight 1
>  via 10.0.8.2, r1-r2-eth3 (vrf default), weight 1
>  via 10.0.9.2, r1-r2-eth4 (vrf default), weight 1
>  via 10.0.10.2, r1-r2-eth5 (vrf default), weight 1
>  Client list: pim(fd 45)

After the r1-r2-th2 and r1-r2-eth4 interfaces are down, the nht entry
for 10.0.5.1 is kept unchanged.

The problem is related to nexthop groups. When using an NHG_ADD to
refresh the BGP route, then NHT entry is not refreshed, and PIM is not
informed about the change.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>