zebra: add list of prefixes attached to protocol nhg
The list of prefixes that use a given nexthop group is not available,
and as consequence, when an event happens on a nexthop group, there
is need to parse the whole list of tables, to check which prefix is
impacted by the nexthop group.

In each nhg_hash_entry, create an hash list that will contain the
list of prefixes added. That list is updated at each ROUTE_ADD
and ROUTE_DELETE actions, once the dataplane succeeded in the update.

The change only populates those hash lists, which will be used in
the next commit.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>