zebra, lib: add srte message flags to nhg_add zapi
Creating a nexthop-group with a color attribute is not possible.
Using such nexthops would help steer traffic to a path that would
be forged by the SR-TE daemon, instead of using the regular path
calculated by the SPF algorithm of IGPs.

This commit is the first of a series that extends the nexthop-group
service with the necessary color information, and solves how to
convey this information from the protocol daemon to the zebra user.

First of all, the color information is used on routes using regular
nexthops: the ZAPI route message includes a message that tells if a
color attribute is present or not on one of the nexthops used.

Add a 32 bit message attribute in the zapi_nhg structure. This field
is conveyed in the ZAPI message, when NHG_ADD is called.

The CP daemons are unchanged, and will mechanically send the message
value of the programmed nexthop-groups to 0.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>