sharpd: add a color option to watch colored nexthop changes
Testing the color of a nexthop requires to watch for nexthop tracking
events from zebra: it tells whether a path for a colored nexthop exists
or not, and it is up to the CP daemon that owns the nexthop-group to
refresh it.

Add a 'color' option to the 'sharp watch nexthop' command. This call
will trigger the nexthop-group to be re-added in zebra.

The next commit will introduce the topotest test about color.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>