pathd: add srv6 manager control over bsid-ipv6 of sr-te policies
A policy can be defined with whatever binding ipv6 sid address,
whereas this value should be uniquely used by the SR-TE service,
and should own a defined srv6 locator.

Add the ability for The pathd daemon to request SRv6 SID values
to the zebra srv6 SID manager.

Compared with the previous behaviour, the user will have to
configure a locator, and the used binding sid should be in this
locator and be freed (not used by an other CP daemon). If not,
the binding SID will be configured but not usable.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>