lib: add srv6 context support for End.B6.Encap SID
The SID manager should be able to provide SID to SRTE policies.
An explicit SID value should be proposed by the pathd daemon to
ZEBRA, based on an usage context.

The usage context is defined by the srv6_sid_ctx structure, and
should help an operator to know the reason why a SID has been
allocated.

For SRTE policies, the usage is a binding SID IPv6 attached to
a SRTE policy. The sid context should be based on the color and
the nexthop of the policy. Propose to extend the srv6_sid_ctx
structure by adding the 'color' attribute, as the ipv6 nexthop
is already present.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Acked-by: Dmytro Shytyi <dmytro.shytyi@6wind.com>