bgpd: add a second hash-list for bgp nexthop groups
Introducing a nexthop-group hierarchy adds complexity on failover
events: the child nexthop-group representation of a nexthop may become
unused, with the following implications:
- the parent(s) nexthop-group may be detached from the unused nexthop:
either the hashes are re-calculated due to the changing number of
childs, or they are detached for an alternate parent nexthop-group.
- that child nexthop-group may be removed.

By using a single hash-list, doing all those operations is not possible
in a frr_each_safe() walk.

Handle this issue by introducing a second hash-list, that will be used
wherever walk operations that need to modify nexthop-groups are
necessary.

That second hash-list will be ordered by identifier.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>