bgpd: add nexthop group framework for l3 routes.
On a BGP device handling a lot of routes with multiple gateways,
if a failover happens, then BGP takes time to syncronise its
installed routes. This is especially true for border routers
that may receive multiple full routes from tier-1 operators.

This takes time because the ZAPI interface between BGP and ZEBRA,
that is in charge of transmitting routes add/updates/removal has
to process a lot of route messages during the transition time.

To reduce the number of messages at ZAPI level, a series of
patches is proposed:
- by splitting the route information in two, and separating the
prefix information from the nexthop information. Two different
messages will be issued
- by avoiding sending route updates when only the identified
nexthop has been changed.

The following changes are extending the BGP Nexthop-group libary,
to use nexthop-groups to L3 routes (today, only EVPN routes could
use nexthop-groups).
A BGP nexthop-group library is created to host the nexthop group
identifier in a new bgp_nhg_cache structure. It reuses the RECURSION,
IBGP and SRTE flags defined in lib/nexthop_group.h.

This commit is only a technical step, and does not use the nexthop group
extension.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>