bgpd, lib: generate aspath json string on demand
When using the 'show bgp ipv4 json detail' against a full route,
the BGP memory consumed increases a lot, and is never restored.

Before the show:
> root@dut-orwell-nianticvf:~# ps -aux | grep bgpd
> root       12263 10.8 10.1 1213920 1033452 ?     Ssl  10:43   1:42 /usr/bin/bgpd -A 127.0.0.1 -M snmp -M rpki -M bmp

After the show:
> root@dut-orwell-nianticvf:~# ps -aux | grep bgpd
> root       11772  4.9 19.3 2150668 1970548 ?     Ssl  08:59   2:09 /usr/bin/bgpd -A 127.0.0.1 -M snmp -M rpki -M bmp

This is not a memory leak, since applying multiple times the
command does not increase the memory.
What happens is that some internal structures of BGP store
a json structure: AsPath, Communities, Large Communities.

This series of commits intends to do the following for the
following objects: ASPaths, communities and large communities:
- replace the usage of the internal json structure by a dynamically
built json object.
- remove the no more used json object referenced in the BGP internals.

That first commit addresses the case with ASpaths: instead of
relying on the internal json structure, the json aspath
objects are allocated and freed at usage.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>