zebra&bgpd : Add PIC support in the srv6 VPN scenario.
This PR implements the separation of nh_srv6-related information, which was originally stored in the nexthop, in the srv6-vpn scenario. It generates a new pic nexthop that contains only the next-hop forwarding information. Meanwhile, the original nexthop containing nh_srv6 is used as pic context information and indexed separately.

Signed-off-by: hanyu.zly&freddy <hanyu.zly@alibaba-inc.com>