isisd: add srv6_locator param to lsp_iterate_ip_reach_cb
When parsing ipv6 reach or locator TLV, the sub-tlvs options
reated to end sid information need to be readable.

The isis_lsp_iterate_ip_reach() function does not have access
to the end sid subtlvs, as this info is not in ip reachs. Add
an extra parameter option to the lsp_ip_reach_iter_cb callback
that gives access to the srv6_locator.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>