pathd: remove srte route when sid allocation fails
Enabling the control of the binding-ipv6-sid value on
an already configured srte policy does not remove the
srte route.

With the below configuration,

> ipv6 route 2001:db8:f000:2::/64 ntfp1
> segment-routing
>  traffic-eng
>   segment-list srv6
>    index 0 ipv6-address 2001:db8:f000:2::
>    index 1 ipv6-address 2001:db8:f000:3::
>   exit
>   policy color 1 endpoint 2001:db8:f000:5::
>    name srv6
>    srv6-binding-sid fc10:0:2::128
>    candidate-path preference 1 name srv6 explicit segment-list srv6
>   exit

if the user enabled binding-ipv6-sid control:

> segment-routing
>  traffic-eng
>   use-srv6-sid-manager

Then the ipv6 route is still present:
> # show ipv6 route
> [..]
> p>* fc10:0:2::128/128 [50/0] is directly connected, ntfp1, seg6local End.B6.Encap nh6 fc10:0:2::128, seg6 2001:db8:f000:2::,2001:db8:
f000:3::, weight 1, 01:46:45

Actually, when SID allocation fails, nothing is done to remove the
previously configured srte route. Remove the route
if already installed.

PR: 91829
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Signed-off-by: Dmytro Shytyi <dmytro.shytyi@6wind.com>
Acked-by: Philippe Guibert <philippe.guibert@6wind.com>