pathd: reinstall srte route when control is disabled back
Following scenario of previous commit, if the user wants
to disable the bsid control, the srte route is not added back.

There is no trigger to reinstall srte route of a valid policy
when the 'use-srv6-sid-manager' command is disabled.

Add a trigger to to install the BSID IPv6 route.

PR: 91829
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Signed-off-by: Dmytro Shytyi <dmytro.shytyi@6wind.com>
Acked-by: Philippe Guibert <philippe.guibert@6wind.com>