bgpd: fix remove too verbose error message
When setting up a RR EVPN setup, the below error message is displayed
for each received BGP update. It becomes flooding when many BGP updates
are received.

> 2025/03/11 14:38:24.959177 BGP: [J51AF-GQ2HJ][EC 33554468] vrf import rt lookup - evpn instance not created yet
> 2025/03/11 14:38:24.959188 BGP: [J51AF-GQ2HJ][EC 33554468] vrf import rt lookup - evpn instance not created yet

The message is just here to inform that there is no BGP backbone
defined, which is normal under a RR configuration. Setting up a RR
EVPN configuration should not lead to such message. Fix this by removing
the log messages.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>