zebra: modify backwalk nhe to handle recursive nhe using IGP
The previous commit ("zebra: add pic support for Kernel based
dataplane") introduces a backwalk algorithm that does not work if the
NHG used by SRv6/MPLS routes recurse over IGP routes.

For instance, the below output shows NHE 81 uses a recursive nexthop.
Also, the AFI of the NHG is not defined.

> r1# show ip route vrf vrf1 nexthop-group
> Codes: K - kernel route, C - connected, L - local, S - static,
>        R - RIP, O - OSPF, I - IS-IS, B - BGP, E - EIGRP, N - NHRP,
>        T - Table, v - VNC, V - VNC-Direct, A - Babel, D - SHARP,
>        F - PBR, f - OpenFabric, t - Table-Direct,
>        > - selected route, * - FIB route, q - queued, r - rejected, b - backup
>        t - trapped, o - offload failure
>
> IPv4 unicast VRF vrf1:
> B>  192.0.2.9/32 [200/0] (93) (pic_nh 81) via 192.0.2.5 (vrf default) (recursive), label 500, weight 1, 00:01:44
>   *                                         via 172.31.0.3, r1-eth1 (vrf default), label 16055/500, weight 1, 00:01:44
>   *                                         via 172.31.2.4, r1-eth2 (vrf default), label 16055/500, weight 1, 00:01:44
> r1# show nexthop-group rib 81
> ID: 81 (zebra)
>      RefCnt: 1
>      Uptime: 00:01:48
>      VRF: default(No AFI)
>      Nexthop Count: 1
>      Valid
>      Depends: (72)
>         via 192.0.2.5 (vrf default) (recursive), weight 1
>            via 172.31.0.3, r1-eth1 (vrf default), weight 1
>            via 172.31.2.4, r1-eth2 (vrf default), weight 1
>      PIC Dependents: (93)

When the 192.0.2.5/32 route deletion is detected by zebra RNH, an exact
NHG is looked up that contains an NHG with IPv4 and 192.0.2.5/32.
Unfortunately, the lookup algorithm also includes the comparison of the
AFI value and, the recursed nexthop, and the number of nexthops.

Propose to relax some attributes in the lookup algorithm, so that the
NHG 81 is matched, and consequently, the BGP route is refreshed by ZEBRA
quicker.

The algorithm is a work in progress yet.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>