bgpd: handle sid allocation inside a locator routemap
A unique auto SID is allocated per route-map, per address-family, per
BGP instance. This patch handles when the SID needs to be allocated, how
VPN path is exported with the new SID.
The bgp_zebra_srv6_sid_notify() function is split, a sub-function is
called each time a bgp locator or a bgp route-map locator matches the
notified sid.

Signed-off-by: Dmytro Shytyi <dmytro.shytyi@6wind.com>
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>