bgpd: add srv6_locator string pointer to extra parameter
The locator name must be stored somewhere on each bgp path info.
At the same time, when a route-map is applied to a bgp path, the
bgp_path_info main struture is not passed to route-map, thus preventing
to directly handle the relationship between a given locator and a
route-map.

This commit proposes to extend the attr_extra structure with a string
pointer. The use case for using locator is just for passing locator from
route-map to the real path_info structure.

Signed-off-by: Dmytro Shytyi <dmytro.shytyi@6wind.com>
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>